refactor(init): migrate plugin Init(INethermindApi) hooks to dedicated steps
Replace the per-plugin `Init(INethermindApi)` hook with dedicated `IStep`
classes so plugin initialization joins the runner step graph with explicit
dependencies instead of a single monolithic `InitializePlugins` pass.

- Merge: `InitializeMergePlugin` (dependents: InitializeBlockchain,
  RegisterRpcModules). PoSSwitcher/InvalidChainTracker are no longer eagerly
  resolved here — the DI graph already constructs them (validators inject
  IPoSSwitcher; InvalidChainTracker via OnActivate<IMainProcessingContext>).
- AuRaMerge: `InitializeAuRaMergePlugin` depends on InitializeMergePlugin and
  installs the AuRa tx-filter hook (dependent: InitializeBlockchain).
  AuRaMergePlugin no longer inherits MergePlugin.
- Taiko: `InitializeTaikoPlugin` wires the L1 precompile statics
  (dependent: LoadGenesisBlock).
- StateDiffsWriter: `InitializeStateDiffsWriterPlugin` starts the pruner,
  which constructs DiffsWriterService (dependent: LoadGenesisBlock).
- Optimism: init was pure null-guards; dropped entirely.
- Shutter: init only logged; the log moved into RunShutterP2P.

Merge.Plugin and Merge.AuRa now reference Nethermind.Init.
