+++ src/Nethermind/Nethermind.Shutter/RunShutterP2P.cs
@@ -9,6 +9,7 @@ using Multiformats.Address;
 using Nethermind.Api.Steps;
 using Nethermind.Config;
 using Nethermind.Init.Steps;
+using Nethermind.Logging;
 using Nethermind.Shutter.Config;
 
 namespace Nethermind.Shutter;
@@ -17,10 +18,14 @@ namespace Nethermind.Shutter;
     dependencies: [typeof(InitializeNetwork)],
     dependents: [typeof(InitializeBlockProducer)]
 )]
-public class RunShutterP2P(IShutterConfig shutterConfig, IShutterApi shutterApi, IProcessExitSource exitSource) : IStep
+public class RunShutterP2P(IShutterConfig shutterConfig, IShutterApi shutterApi, IProcessExitSource exitSource, ILogManager logManager) : IStep
 {
+    private readonly ILogger _logger = logManager.GetClassLogger<RunShutterP2P>();
+
     public Task Execute(CancellationToken cancellationToken)
     {
+        if (_logger.IsInfo) _logger.Info("Initializing Shutter plugin.");
+
         IEnumerable<Multiaddress> bootnodeP2PAddresses;
         try
         {
