+++ src/Nethermind/Nethermind.StateDiffsWriter/StateDiffsWriterPlugin.cs
@@ -1,12 +1,8 @@
 // SPDX-FileCopyrightText: 2026 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
-using System.Threading.Tasks;
-using Autofac;
 using Autofac.Core;
-using Nethermind.Api;
 using Nethermind.Api.Extensions;
-using Nethermind.StateDiffsWriter.Service;
 
 namespace Nethermind.StateDiffsWriter;
 
@@ -23,14 +19,4 @@ public class StateDiffsWriterPlugin(IStateDiffsWriterConfig config) : INethermin
     public bool Enabled => config.Enabled;
     public bool MustInitialize => true;
     public IModule Module => new StateDiffsWriterModule();
-
-    public Task Init(INethermindApi nethermindApi)
-    {
-        // Resolve eagerly so the NewHeadBlock subscription attaches before the first head;
-        // container shutdown disposes both singletons, so no explicit teardown is needed.
-        _ = nethermindApi.Context.Resolve<DiffsWriterService>();
-        DiffsPruner pruner = nethermindApi.Context.Resolve<DiffsPruner>();
-        pruner.Start();
-        return Task.CompletedTask;
-    }
 }
